+++ changelog.md
@@ -1,3 +1,6 @@
+- Add 'keypasswd': Allows setting password for pem keyfile (#1676)
+- Bugfix:w Improve warning messages for unexpected frame type error to track origin (#1671)
+- Bugfix: Remove special chars in auto-generated client id (#1673)
 - Bugfix: Websocket returned error 500 and wrote to log, instead of returning 426 (protocol upgrade) #1983 
 - Allow to specify a maximum connection lifetime (per listener). The lifetime can be overwriten by on_register hooks.
 - Improve TLSv1.3 support (Documentation, CLI, Testsuite)